+++ ydb/core/tx/locks/ya.make
@@ -17,8 +17,3 @@ PEERDIR(
 YQL_LAST_ABI_VERSION()
 
 END()
-
-
-RECURSE_FOR_TESTS(
-    ut_range_treap
-)
