+++ ydb/library/range_treap/ya.make
@@ -11,3 +11,7 @@ PEERDIR(
 )
 
 END()
+
+RECURSE_FOR_TESTS(
+    ut_range_treap
+)
